mux_socket: Remove dupped file descriptors
Clean up the I/O model for multiplexed trasnport by not using dupped sockets.

A dummy socket is still used as the Envoy Connection object wants to own one,
but it is not used for anything else than as a globally unique ID for the
connection in the MuxData maps as well as for tracking connection closure.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>