policy: Pause NPDS until worker threads are done with the update
Normally Envoy ACKs xDS updates right after the main thread is done
with the update, while the worker threads may still be applying the
update. This has been observed to lead to incoming traffic initiated
after the ACK to be processed by a worker with stale
configuration. Fix this by pausing the gRPC subscription while the
worker threads are applying the updates.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>